requirements.txt
python-dotenv==1.0.0
opencv-python-headless==4.5.5.64
pytesseract==0.3.9
Pillow==8.4.0
pdf2image==1.16.0
PyMuPDF==1.19.6
langchain==0.0.151
langchain-huggingface==0.0.1
langchain-chroma==0.0.4
